+++ doc/CHANGES.rst
@@ -20,6 +20,10 @@ Future
 * [FIX] add set_defaults in parser to update configuration dict
 * [FIX] one2many remote columns
 * [FIX] load anyblok.init in the unit test
+* [IMP] Add plugins by configuration for
+  - Registry
+  - Migration
+  - get_url
 
 0.8.5 (2016-06-20)
 ------------------